making of eip-3076 (#7894)
