+++ account_manager/Cargo.toml
@@ -19,12 +19,13 @@ eth2_keystore = { workspace = true }
 eth2_network_config = { workspace = true }
 eth2_wallet = { workspace = true }
 eth2_wallet_manager = { path = "../common/eth2_wallet_manager" }
+eip_3076 = { workspace = true }
 filesystem = { workspace = true }
 safe_arith = { workspace = true }
 sensitive_url = { workspace = true }
 serde_json = { workspace = true }
-slashing_protection = { workspace = true }
 slot_clock = { workspace = true }
+slashing_protection = { workspace = true }
 tokio = { workspace = true }
 types = { workspace = true }
 validator_dir = { workspace = true }
