+++ beacon_node/store/Cargo.toml
@@ -17,7 +17,6 @@ ethereum_ssz = { workspace = true }
 ethereum_ssz_derive = { workspace = true }
 itertools = { workspace = true }
 leveldb = { version = "0.8.6", optional = true, default-features = false }
-logging = { workspace = true }
 lru = { workspace = true }
 metrics = { workspace = true }
 parking_lot = { workspace = true }
@@ -30,7 +29,6 @@ state_processing = { workspace = true }
 strum = { workspace = true }
 superstruct = { workspace = true }
 tracing = { workspace = true }
-tracing-subscriber = { workspace = true }
 types = { workspace = true }
 xdelta3 = { workspace = true }
 zstd = { workspace = true }
