+++ common/eth2/Cargo.toml
@@ -16,6 +16,7 @@ eth2_keystore = { workspace = true }
 ethereum_serde_utils = { workspace = true }
 ethereum_ssz = { workspace = true }
 ethereum_ssz_derive = { workspace = true }
+eip_3076 = { workspace = true }
 futures = { workspace = true }
 futures-util = "0.3.8"
 libp2p-identity = { version = "0.2", features = ["peerid"] }
@@ -29,7 +30,6 @@ reqwest-eventsource = "0.5.0"
 sensitive_url = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
-slashing_protection = { workspace = true }
 ssz_types = { workspace = true }
 test_random_derive = { path = "../../common/test_random_derive" }
 types = { workspace = true }
