+++ lighthouse/Cargo.toml
@@ -87,6 +87,7 @@ eth2 = { workspace = true }
 initialized_validators = { workspace = true }
 lighthouse_network = { workspace = true }
 sensitive_url = { workspace = true }
+eip_3076 = { workspace = true }
 slashing_protection = { workspace = true }
 tempfile = { workspace = true }
 validator_dir = { workspace = true }
