+++ testing/web3signer_tests/Cargo.toml
@@ -22,6 +22,7 @@ reqwest = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 serde_yaml = { workspace = true }
+eip_3076 = { workspace = true }
 slashing_protection = { workspace = true }
 slot_clock = { workspace = true }
 task_executor = { workspace = true }
