+++ validator_client/http_api/Cargo.toml
@@ -32,6 +32,7 @@ sensitive_url = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 signing_method = { workspace = true }
+eip_3076 = { workspace = true }
 slashing_protection = { workspace = true }
 slot_clock = { workspace = true }
 sysinfo = { workspace = true }
