+++ validator_client/slashing_protection/src/interchange_test.rs
@@ -1,7 +1,10 @@
 use crate::{
-    SigningRoot, SlashingDatabase,
-    interchange::{Interchange, SignedAttestation, SignedBlock},
+    SlashingDatabase,
     test_utils::{DEFAULT_GENESIS_VALIDATORS_ROOT, pubkey},
+};
+use eip_3076::{
+    SigningRoot,
+    interchange::{Interchange, SignedAttestation, SignedBlock},
 };
 use serde::{Deserialize, Serialize};
 use std::collections::HashSet;
